requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.39
python-dotenv==0.21.0
gunicorn==20.1.0
psycopg2-binary==2.9.5